requirement.txt
absl-py==0.9.0
alabaster==0.7.12
asgiref==3.3.1
asn1crypto==1.3.0
astor==0.8.1
astroid==2.3.3
attrs==19.3.0
Babel==2.8.0
backcall==0.2.0
beautifulsoup4==4.9.1
bleach==3.1.5
cachetools==4.1.0
certifi==2020.6.20
cffi==1.13.2
chardet==3.0.4
click==7.1.2
cloudpickle==1.2.2
colorama==0.4.3
cryptography==2.8
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
Django==3.1.7
docutils==0.16
entrypoints==0.3
et-xmlfile==1.0.1
Flask==1.1.2
future==0.18.2
gast==0.2.2
google-auth==1.18.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.30.0
gTTS==2.2.2
h5py==2.10.0
idna==2.9
image==1.5.33
imagesize==1.2.0
importlib-metadata==1.6.1
imutils==0.5.3
inflect==4.1.0
ipykernel==5.3.0
ipython==7.15.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
jaraco.itertools==5.0.0
jdcal==1.4.1
jedi==0.17.1
Jinja2==2.11.2
joblib==0.15.1
json5==0.9.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-core==4.6.3
jupyterlab==1.2.6
jupyterlab-server==1.0.6
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keyring==21.1.0
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
lxml==4.5.1
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.1.1
mccabe==0.6.1
mistune==0.8.4
mkl-service==2.3.0
more-itertools==8.2.0
nbconvert==5.6.1
nbformat==5.0.7
notebook==6.0.3
numpy==1.19.0
numpydoc==0.9.2
oauthlib==3.1.0
olefile==0.46
opencv-python==4.2.0.32
openpyxl==3.0.3
opt-einsum==3.2.1
packaging==20.4
pandas==0.25.2
pandocfilters==1.4.2
parso==0.7.0
pickleshare==0.7.5
Pillow==6.2.1
playsound==1.2.2
prometheus-client==0.8.0
prompt-toolkit==3.0.5
protobuf==3.12.2
psutil==5.6.7
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.48
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.1
pyglet==1.5.3
Pygments==2.6.1
pylint==2.4.4
PyMsgBox==1.0.7
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyrsistent==0.16.0
PyScreeze==0.1.26
PySocks==1.7.1
python-dateutil==2.8.1
PyTweening==1.0.3
pytz==2020.1
pywin32==228
pywin32-ctypes==0.2.0
pywinpty==0.5.7
pyzmq==19.0.1
QtAwesome==0.6.1
qtconsole==4.7.5
QtPy==1.9.0
requests==2.24.0
requests-oauthlib==1.3.0
rope==0.16.0
rsa==4.6
scikit-learn==0.21.3
scipy==1.5.0
seaborn==0.9.0
Send2Trash==1.5.0
six==1.15.0
snowballstemmer==2.0.0
soupsieve==2.0.1
Sphinx==2.3.1
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
spyder==3.3.6
spyder-kernels==0.5.2
sqlparse==0.4.1
tensorboard==2.0.2
tensorflow==2.0.0
tensorflow-estimator==2.0.1
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
tornado==6.0.4
traitlets==4.3.3
typed-ast==1.4.1
urllib3==1.25.9
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
win-inet-pton==1.1.0
wincertstore==0.2
wrapt==1.12.1
xlrd==1.2.0
zipp==3.1.0
